Organize code in package for new feature
