+++ src/helpers/test_helpers.jl
@@ -21,7 +21,12 @@ end
 
 
 """
-    getzerobase(T_type=Float32, G=218, T_periods=120, startdate=Date(2001,1), baseindex=100*one(T_type))
+    getzerobase(; 
+        [T_type = Float32, 
+        G = 218, 
+        T_periods = 120, 
+        startdate = Date(2001,1), 
+        baseindex = 100*one(T_type)])
 Función para obtener base de tipo `VarCPIBase` con variaciones intermensuales
 iguales a cero.
 """
@@ -32,19 +37,47 @@ function getzerobase(T_type=Float32, G=218, T_periods=120, startdate=Date(2001,1
     VarCPIBase(vmat, w, dates, baseindex)
 end
 
+function getzerobase(;
+    T_type = Float32, 
+    G = 218, 
+    T_periods = 120, 
+    startdate = Date(2001,1), 
+    baseindex = 100*one(T_type))
+
+    getzerobase(T_type, G, T_periods, startdate, baseindex)
+end
+
+
 """
-    getzerocountryst(T_type=Float32)
+    getzerocountryst(T_type = Float32)
 Obtiene un `UniformCountryStructure` cuyas variaciones intermensuales son todas
 iguales a cero en la configuración de períodos del IPC de Guatemala.
 """
-function getzerocountryst(T_type=Float32)
+function getzerocountryst(T_type = Float32)
     gt00 = getzerobase(T_type, 218, 120, Date(2001, 1))
     gt10 = getzerobase(T_type, 279, 120, Date(2011, 1))
     UniformCountryStructure(gt00, gt10)
 end
 
 """
-    getrandombase(T_type=Float32, G=218, T_periods=120, startdate=Date(2001,1), baseindex=100*one(T_type))
+    getrandomcountryst(T_type = Float32)
+Obtiene un `UniformCountryStructure` cuyas variaciones intermensuales son todas
+aleatorias en la configuración de períodos del IPC de Guatemala.
+"""
+function getrandomcountryst(T_type = Float32)
+    gt00 = getrandombase(T_type, 218, 120, Date(2001, 1))
+    gt10 = getrandombase(T_type, 279, 120, Date(2011, 1))
+    UniformCountryStructure(gt00, gt10)
+end
+
+
+"""
+    getrandombase(; 
+        [T_type = Float32, 
+        G = 218, 
+        T_periods = 120, 
+        startdate = Date(2001,1), 
+        baseindex = 100*one(T_type)])
 Función para obtener una base de tipo `VarCPIBase` con variaciones
 intermensuales aleatorias.
 """
@@ -53,4 +86,14 @@ function getrandombase(T_type=Float32, G=218, T_periods=120, startdate=Date(2001
     w = getrandomweights(T_type, G)
     dates = getbasedates(vmat, startdate)
     VarCPIBase(vmat, w, dates, baseindex)
+end
+
+function getrandombase(;
+    T_type = Float32, 
+    G = 218, 
+    T_periods = 120, 
+    startdate = Date(2001,1), 
+    baseindex = 100*one(T_type))
+
+    getrandombase(T_type, G, T_periods, startdate, baseindex)
 end
